requirements.txt
langchain
langchainhub
bs4
tiktoken
openai
boto3==1.34.37
langchain_community
chromadb
pinecone
awscli
pypdf
faiss-cpu
langchain_aws
tiktoken
langchain-openai
openai
fastapi
uvicorn
python-dotenv
python-multipart
-e .